Remove feast import from featurebyte/query_graph/node/schema.py
